+++ lib/modules/auth/auth_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: auth
-// Updated at: 2026-05-18 17:37:23
-// Commit: feat(auth): implement shimmer loading skeleton in auth
+// Updated at: 2026-05-20 16:29:52
+// Commit: chore(auth): update dependencies and cleanup unused imports in auth
 
 class authComponentState {
-  final String commitRef = '1263';
-  final String lastMessage = 'feat(auth): implement shimmer loading skeleton in auth';
-  final DateTime timestamp = DateTime.parse('2026-05-18T17:37:23+05:00');
+  final String commitRef = '1297';
+  final String lastMessage = 'chore(auth): update dependencies and cleanup unused imports in auth';
+  final DateTime timestamp = DateTime.parse('2026-05-20T16:29:52+05:00');
 }